admin/requirements.txt
# Upgrade pip : pip install -i https://pypi.tuna.tsinghua.edu.cn/simple --upgrade pip
# Install with: pip install -i https://pypi.tuna.tsinghua.edu.cn/simple -r requirements.txt

# Wheel
wheel >= 0.30.0

# Django
django >= 2.0.0
django-cors-headers >= 2.1.0
django-daterange-filter >= 1.3.0
django-extensions >= 1.9.7
django-import-export >= 0.5.1
django-tinymce >= 2.6.0
djangorestframework >= 3.7.3
djangorestframework-jwt >= 1.11.0


mysqlclient >= 1.3.12
requests >= 2.18.4
bs4 >= 0.0.1
lxml >= 4.1.1
jieba >= 0.39
python-dateutil >= 2.7.2

baidu-aip >= 2.2.0.0